Migration to acorn 6 new plugin API
In acorn 6, the API for plugins changed and these changes conform to the new API.

All tests are still be passing without any changes!
