+++ package.json
@@ -25,7 +25,8 @@
     "tag": "git tag v$npm_package_version"
   },
   "dependencies": {
-    "acorn": "^5.0.0"
+    "acorn": "^6.0.1",
+    "acorn-walk": "^6.0.1"
   },
   "devDependencies": {
     "babel-cli": "^6.18.0",
